+++ NEWS.md
@@ -5,6 +5,7 @@
 - Decouple primary keys and foreign keys from index creation
 - Allow ordering of indexes
 - Order of fields in index is conserved (not sorted by alphabetic order aymore)
+- Improved display of fields and comments in data model plot
 
 <!----------------------------------------------------------------------------->
 <!----------------------------------------------------------------------------->